set main screen as the most recent screen entered
We now keep track of all the outputs the surface is using and in which order it
entered them to select the most recently entered output as the main one.

Spanning multiple monitors at once and selectively moving out of a few of them
in a different order is supported: every time the surface enters or exits one,
we update our reference to be the most recently entered output still in use.

This also means hotplugging is supported.